+++ tests/Support/Stubs/Forms/CreateUserFromRequest.php
@@ -23,7 +23,7 @@ class CreateUserFromRequest extends FormRequest
             'roles.*.id'         => 'required|min:1|max:255',
             'roles.*.name'       => 'min:1|max:255',
             'permissions'        => 'array',
-            'permissions.*.id'   => 'required|min:1|max:255',
+            'permissions.*.id'   => ['required', 'min:1', 'max:255'],
             'permissions.*.name' => 'min:1|max:255',
             'spare'              => 'array',
             'spare.*'            => 'min:1|max:255',